Se agregoa a la pagina principal el metodo para llamar al metodo ActualizarFactura y se corrigio un error de variable

+++ Components/Data/ServicioFactura.cs
@@ -72,7 +72,7 @@ namespace hombresYMujeresFacturan.Components.Data
             }
         }
 
-        public async Task ActualizarFactura(int identificador,string fecha, string nombrecliente, string articulos, int total)
+        public async Task ActualizarFactura(int identificador,string fecha, string nombrecliente, string articulos, decimal total)
         {
             string ruta = "mibase.db";
             using var conexion = new SqliteConnection($"DataSource={ruta}");
